Replace beaker-puppet_install_helper with fixtures
This copies the approach taken by litmus and builds every module in
fixtures. Those are then installed on the target system.

This has the benefit that it's much faster since there's no more
reliance on the Puppet forge. It's also more flexible since you can use
git versions in fixtures.