Move to beaker_puppet_helpers
This ends up dropping beaker-pe, beaker-puppet,
beaker-puppet_install_helper and beaker-module_install_helper.
This drops support for Puppet Enterprise installations and development
builds. However, it does drop the dependency on beaker-pe and greatly
simplifies the dependency tree. It also saves maintenance on multiple
gems.

Another difference is that it now defaults to the puppet collections
(instead of puppet6) which means the latest stable release.

Installation of modules now happens by creating a module release (using
puppet-modulebuilder) and running puppet module install on that. This is
a lot faster than the old method because it happens in a single
transaction.

The new puppet_package_name method can return the correct non-AIO
package name. This allows testing on platforms where there is no AIO
package available.