puppetlabs_spec_helper: require 4.x and newer
version 3 defines the beaker rake task. beaker-rspec defines that as
well, and that's the one we want to use. We need to ensure that we load
puppetlabs_spec_helper in version 4 (which doesn't have the beaker rake
task). Otherwise we end up with two tasks. Calling it will run both
which in turn will run all tests twice.