Make puppet agent installation configureable
When `BEAKER_PROVISION_PUPPET` is set to `yes` (which is the default), `puppet` will be installed. This can be used in combination with `BEAKER_PROVISON`. Set `BEAKER_PROVSION=yes` and `BEAKER_PROVISON_PUPPET=no` to spin up a new instance but assume puppet is already installed. This is helpful if you test with your own images or in isolated environments where the instance cannot reach the puppet.com RPM/Deb mirrors.